move static_test
