+++ CMakeLists.txt
@@ -11,4 +11,5 @@ add_executable(${CMAKE_PROJECT_NAME} main.cpp main.cpp)
 target_include_directories(${CMAKE_PROJECT_NAME} PRIVATE
         ${CMAKE_CURRENT_SOURCE_DIR}/include)
 
+add_subdirectory(static_test)
 add_subdirectory(test)
