+++ static_test/concepts/geometry/point_concept.cpp
