+++ static_test/concepts/graphics/color_concept.cpp
