+++ static_test/concepts/ui/drawable.cpp
@@ -1,7 +1,7 @@
 #include <usagi/concepts/ui/drawable.hpp>
 
-#include "../../mocks/draw_context.hpp"
-#include "../../mocks/drawable.hpp"
+#include "../../../test/mocks/draw_context.hpp"
+#include "../../../test/mocks/drawable.hpp"
 
 namespace {
 static_assert(usagi::concepts::ui::drawable<Drawable<float, DrawContext>>);
