+++ static_test/concepts/ui/gestural.cpp
@@ -1,6 +1,6 @@
 #include <usagi/concepts/ui/gestural.hpp>
 
-#include "../../mocks/gestural.hpp"
+#include "../../../test/mocks/gestural.hpp"
 
 namespace {
 static_assert(usagi::concepts::ui::gestural<Gestural<float>>);
