+++ static_test/concepts/ui/viewable.cpp
@@ -1,6 +1,6 @@
 #include <usagi/concepts/ui/viewable.hpp>
 
-#include "../../mocks/viewable.hpp"
+#include "../../../test/mocks/viewable.hpp"
 
 namespace {
 static_assert(usagi::concepts::ui::viewable<Viewable<int>>);
