+++ static_test/main.cpp
@@ -1,0 +1,4 @@
+int main(int argc, char **argv)
+{
+  return 0;
+}
