+++ static_test/ui/base_view.cpp
@@ -1,4 +1,3 @@
-#include <usagi/concepts/ui/hierarchy.hpp>
 #include <usagi/ui/detail/base_view_hierarchy.hpp>
 #include <usagi/ui/view.hpp>
