+++ static_test/variable/variable_traits.cpp
