+++ test/geometry/point/point.cpp
@@ -1,12 +1,6 @@
 #include <gtest/gtest.h>
-#include <usagi/concepts/geometry/point_concept.hpp>
 #include <usagi/geometry/point/point.hpp>
 #include <usagi/geometry/size/size.hpp>
-
-namespace {
-static_assert(usagi::concepts::geometry::point_concept<usagi::geometry::point<int>>);
-static_assert(usagi::concepts::geometry::point_concept<usagi::geometry::variable_point<int>>);
-}
 
 TEST(PointTest, ConstructorCase) {
   {
